requirements.txt
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-mistralai>=0.2.0
openai>=1.50.0
requests>=2.31.0
python-dotenv>=1.0.0
streamlit>=1.38.0
